adding custom-output-path-app tests
