+++ test-packages/custom-output-path-app/README.md
@@ -1,4 +1,4 @@
-# custom-finger-print-app
+# custom-output-path-app
 
 This README outlines the details of collaborating on this Ember application.
 A short introduction of this app could easily go here.
@@ -15,7 +15,7 @@ You will need the following things properly installed on your computer.
 ## Installation
 
 * `git clone <repository-url>` this repository
-* `cd custom-finger-print-app`
+* `cd custom-output-path-app`
 * `npm install`
 
 ## Running / Development
