+++ test-packages/custom-output-path-app/app/index.html
@@ -10,7 +10,7 @@
     {{content-for "head"}}
 
     <link integrity="" rel="stylesheet" href="{{rootURL}}assets/vendor.css">
-    <link integrity="" rel="stylesheet" href="{{rootURL}}assets/custom-finger-print-app.css">
+    <link integrity="" rel="stylesheet" href="{{rootURL}}assets/custom-output-path-app.css">
 
     {{content-for "head-footer"}}
   </head>
@@ -18,7 +18,7 @@
     {{content-for "body"}}
 
     <script src="{{rootURL}}assets/vendor.js"></script>
-    <script src="{{rootURL}}assets/custom-finger-print-app.js"></script>
+    <script src="{{rootURL}}assets/custom-output-path-app.js"></script>
 
     {{content-for "body-footer"}}
   </body>
