--- test-packages/custom-output-path-app/config/fastboot.js
@@ -1,7 +1,0 @@
-module.exports = function (environment) {
-  return {
-    sandboxGlobals: {
-      myGlobal: 'My Global'
-    }
-  };
-}
