+++ test-packages/custom-output-path-app/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "custom-finger-print-app",
+  "name": "custom-output-path-app",
   "version": "0.0.0",
   "private": true,
-  "description": "Small description for custom-finger-print-app goes here",
+  "description": "Small description for custom-output-path-app goes here",
   "repository": "",
   "license": "MIT",
   "author": "",
